dualgpuopt/requirements_legal.txt
# Core dependencies
peft>=0.4.0
transformers>=4.31.0
accelerate>=0.21.0
torch>=2.0.1
sentencepiece>=0.1.99
datasets>=2.13.0
pyyaml>=6.0
tqdm>=4.65.0

# RAG dependencies
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU support
sentence-transformers>=2.2.2
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Evaluation dependencies
pytest>=7.3.1
numpy>=1.24.3

# API dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
prometheus-client>=0.17.0  # Optional

# HTML processing
beautifulsoup4>=4.12.0
lxml>=4.9.2
